% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additional.R, R/functions.R, R/qc.R
\name{Changepoint}
\alias{Changepoint}
\alias{Changepoint}
\alias{Changepoint}
\title{Changepoint Detection (both uptrend and downtrend)}
\usage{
Changepoint(inset, eval.date, pen.drop = 0.2, zerorate = (3/4), n.mos = 3)

Changepoint(inset, eval.date, pen.drop = 0.2, zerorate = (3/4), n.mos = 3)

Changepoint(inset, eval.date, pen.drop = 0.2, zerorate = (3/4), n.mos = 3)
}
\description{
Returns:
Logical of significant trend found

Returns:
Logical of significant trend found

Returns:
Logical of significant trend found
}
\details{
Parameter List:

inset
-----
Input data frame with the following columns:
dt.ym = Date format with each row being a unique month
dpmo = Number of complaints

eval.date
---------
Which Date-format date in inset$dt.ym to run algorithm on

pen.drop
-----
The minimum proportion drop in penalty to qualify a set of changepoints

zerorate
--------
Maximum proportion of months containing zeroes for this algorithm to run

n.mos
--------
Number of months back from the eval.date to detect a changepoint

Parameter List:

inset
-----
Input data frame with the following columns:
dt.ym = Date format with each row being a unique month
dpmo = Number of complaints

eval.date
---------
Which Date-format date in inset$dt.ym to run algorithm on

pen.drop
-----
The minimum proportion drop in penalty to qualify a set of changepoints

zerorate
--------
Maximum proportion of months containing zeroes for this algorithm to run

n.mos
--------
Number of months back from the eval.date to detect a changepoint

Parameter List:

inset
-----
Input data frame with the following columns:
dt.ym = Date format with each row being a unique month
dpmo = Number of complaints

eval.date
---------
Which Date-format date in inset$dt.ym to run algorithm on

pen.drop
-----
The minimum proportion drop in penalty to qualify a set of changepoints

zerorate
--------
Maximum proportion of months containing zeroes for this algorithm to run

n.mos
--------
Number of months back from the eval.date to detect a changepoint
}
