% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{Cox.Stuart}
\alias{Cox.Stuart}
\title{Cox-Stuart Signal}
\usage{
Cox.Stuart(inset, eval.date, pcrit = 0.05, zerorate = (3/4))
}
\description{
Returns:
Logical of significant trend found
}
\details{
Parameter List:

inset
-----
Input data frame with the following columns:
dt.ym = Date format with each row being a unique month
dpmo = Calculated complaint rate

eval.date
---------
Which Date-format date in inset$dt.ym to run algorithm on

pcrit
-----
Critical p-value to evaluate test

zerorate
--------
Maximum proportion of months containing zeroes for this algorithm to run
}
