% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additional.R, R/functions.R, R/qc.R
\name{EWMA}
\alias{EWMA}
\alias{EWMA}
\alias{EWMA}
\title{EWMA Signals - 4 Western Electric Rules}
\usage{
EWMA(inset, eval.date, lambda = 0.5, zerorate = (3/4))

EWMA(inset, eval.date, lambda = 0.5, zerorate = (3/4))

EWMA(inset, eval.date, lambda = 0.5, zerorate = (3/4))
}
\description{
Returns:
Vector of 4 with a logical of whether the Zone Rules 1-4 were met

Returns:
Vector of 4 with a logical of whether the Zone Rules 1-4 were met

Returns:
Vector of 4 with a logical of whether the Zone Rules 1-4 were met
}
\details{
Parameter List:

inset
-----
Input data frame with the following columns:
dt.ym = Date format with each row being a unique month
dpmo = Calculated complaint rate

eval.date
---------
Which Date-format date in inset$dt.ym to run algorithm on

lambda
-----
Weighting variable

zerorate
--------
Maximum proportion of months containing zeroes for this algorithm to run

Parameter List:

inset
-----
Input data frame with the following columns:
dt.ym = Date format with each row being a unique month
dpmo = Calculated complaint rate

eval.date
---------
Which Date-format date in inset$dt.ym to run algorithm on

lambda
-----
Weighting variable

zerorate
--------
Maximum proportion of months containing zeroes for this algorithm to run

Parameter List:

inset
-----
Input data frame with the following columns:
dt.ym = Date format with each row being a unique month
dpmo = Calculated complaint rate

eval.date
---------
Which Date-format date in inset$dt.ym to run algorithm on

lambda
-----
Weighting variable

zerorate
--------
Maximum proportion of months containing zeroes for this algorithm to run
}
