% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additional.R, R/functions.R
\name{Fwd.Momentum}
\alias{Fwd.Momentum}
\alias{Fwd.Momentum}
\title{Forward-Momentum Signal}
\usage{
Fwd.Momentum(inset, eval.date, pcrit = 0.05, zerorate = (3/4))

Fwd.Momentum(inset, eval.date, pcrit = 0.05, zerorate = (3/4))
}
\description{
Returns:
Logical of significant trend found

Returns:
Logical of significant trend found
}
\details{
Parameter List:

inset
-----
Input data frame with the following columns:
dt.ym = Date format with each row being a unique month
count = Number of complaints

eval.date
---------
Which Date-format date in inset$dt.ym to run algorithm on

pcrit
-----
Critical p-value to evaluate test

zerorate
--------
Maximum proportion of months containing zeroes for this algorithm to run

Parameter List:

inset
-----
Input data frame with the following columns:
dt.ym = Date format with each row being a unique month
count = Number of complaints

eval.date
---------
Which Date-format date in inset$dt.ym to run algorithm on

pcrit
-----
Critical p-value to evaluate test

zerorate
--------
Maximum proportion of months containing zeroes for this algorithm to run
}
