% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mds_ts}
\alias{mds_ts}
\title{Sample List of \code{mds_ts} Time Series}
\format{A list with 3 elements, with each element being a data frame of class}
\source{
\code{maude} and \code{sales} datasets, run through \code{mds}
package functions.
}
\usage{
mds_ts
}
\description{
A list of three \code{mds_ts} time series generated using the \code{mds}
package function \code{time_series()}. Its underlying data were queried from
the [FDA MAUDE API](https://open.fda.gov/api/). Furthermore, a simulated
exposure dataset was generated to provide exposure data.
}
\keyword{datasets}
