% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{Poisson.Rare}
\alias{Poisson.Rare}
\title{Poisson for Rare Events}
\usage{
Poisson.Rare(inset, eval.date, zerorate = (2/3), prate = 0.2,
  pcrit = 0.05)
}
\description{
Returns:
Logical of significant trend found
}
\details{
Parameter List:

inset
-----
Input data frame with the following columns:
dt.ym = Date format with each row being a unique month
count = Absolute number of complaints

eval.date
---------
Which Date-format date in inset$dt.ym to run algorithm on

zerorate
--------
Minimum proportion of months containing zeroes for this algorithm to run

prate
-----
Hypothesized Poisson monthly rate at which the test is perform (null vs. greater)

pcrit
-----
Critical p-value for test
}
