% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additional.R
\name{poisson_rare}
\alias{poisson_rare}
\title{Poisson for Rare Events}
\usage{
poisson_rare(df, ...)
}
\arguments{
\item{df}{Required input data frame of class \code{mds_ts} or, for generic
usage, any data frame with the following columns:
\describe{
  \item{time}{Unique times of class \code{Date}}
  \item{event}{Either the event count or rate of class \code{numeric}}
}}

\item{ts_event}{Required if \code{df} is of class \code{mds_ts}. Named string
indicating the variable corresponding to the event count or rate. Rate must
be calculated in a separate column in \code{df} as it is not calculated by
default. The name of the string is an English description of what was
analyzed.

Default: \code{"nA"} corresponding to the event count column in \code{mds_ts}
objects. Name is generated from \code{mds_ts} metadata.

Example: \code{stats::setNames("rate", "Rate of Bone Filler Events in Canada")}}

\item{analysis_of}{Optional string indicating the English description of what
was analyzed. If specified, this will override the name of the
\code{ts_event} string parameter.

Default: \code{NA} indicates no English description.

Example: \code{"Rate of bone cement leakage"}}

\item{eval_period}{Optional positive integer indicating the number of unique
times counting in reverse chronological order to assess.

Default: \code{NULL} considers all times in \code{df}.}

\item{zero_rate}{Minimum proportion of \code{event}s in \code{df}
(constrained by \code{eval_period}) containing zeroes for this algorithm to
run.

Default: \code{2/3} requires a minimum of 2/3 zeros in \code{event}s in
\code{df}.}

\item{p_rate}{Hypothesized Poisson rate parameter null value at which the
Poisson test is performed (null vs. greater). See details for more.

Default: \code{0.2}}

\item{p_crit}{Critical p-value for the Poisson test..

Default: \code{0.05}}
}
\value{
A named list of class \code{mdsstat_test} object, as follows:
\describe{
  \item{test_name}{Name of the test run}
  \item{analysis_of}{English description of what was analyzed}
  \item{status}{Named boolean of whether the test was run. The name contains
  the run status.}
  \item{result}{A standardized list of test run results: \code{statistic}
  for the test statistic, \code{ll95} and \code{ul95} for the 95%
  confidence bounds, \code{p} for the p-value, \code{signal} status, and
  \code{signal_threshold}.}
  \item{params}{The test parameters}
  \item{data}{The data on which the test was run}
}
}
\description{
Test on rare events using an exact test on the Poisson distribution rate
parameter (\code{stats::poisson.test()}).
}
\details{
\code{p_rate} default of \code{0.2} is a suggested null value for
the Poisson rate parameter. However this value is highly advised to be set
based on known priors and/or your specific application.
}
\examples{
# Basic Example
data <- data.frame(time=c(1:8), event=c(rep(0, 6), rpois(2, 4)))
a1 <- poisson_rare(data)
# Example using an mds_ts object
a2 <- poisson_rare(mds_ts[[1]])
# Example using a derived rate as the "event"
data <- mds_ts[[1]]
data$rate <- ifelse(is.na(data$nA), 0, data$nA) / data$exposure
a3 <- poisson_rare(data, "rate")
}
