% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{Shewart}
\alias{Shewart}
\title{Shewart Signals - 4 Western Electric Rules}
\usage{
Shewart(inset, eval.date, n.mos = 12, zerorate = (1/3))
}
\description{
Returns:
Vector of 4 with a logical of whether the Zone Rules 1-4 were met
}
\details{
Parameter List:

inset
-----
Input data frame with the following columns:
dt.ym = Date format with each row being a unique month
dpmo = Calculated complaint rate

eval.date
---------
Which Date-format date in inset$dt.ym to run algorithm on

n.mos
-----
Number of months to look back when calculating algorithm

zerorate
--------
Maximum proportion of months containing zeroes for this algorithm to run
}
