% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{test2row}
\alias{test2row}
\title{Coerce mdsstat Test to 1-Row Data Frame}
\usage{
test2row(df)
}
\arguments{
\item{df}{Required input object of class \code{mdsstat_test}}
}
\value{
1-row data frame (class \code{mdsstat_df}) summarizing the test.
}
\description{
Coerce an mdsstat test (class \code{mdsstat_test}) to a 1-row data frame.
}
\examples{
test2row(prr(mds_ts[[3]]))
}
