% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additional.R, R/functions.R, R/qc.R
\name{Trend.Signals}
\alias{Trend.Signals}
\alias{Trend.Signals}
\alias{Trend.Signals}
\title{Return All Trending Signals}
\usage{
Trend.Signals(this, mo, warnings = F)

Trend.Signals(this, mo, warnings = F)

Trend.Signals(this, mo, warnings = F)
}
\description{
Example: Trend.Signals(dataset, as.Date("2015-10-01"))

Example: Trend.Signals(dataset, as.Date("2015-10-01"))

Example: Trend.Signals(dataset, as.Date("2015-10-01"))
}
\details{
Returns:
Logical vector of trends by algorithm

Parameter List:

this
-----
Input data frame with the following columns:
dt.ym = Date format with each row being a unique month
dpmo = Calculated complaint rate
count = Number of complaints

mo
---------
Which Date-format date from this$dt.ym to run algorithms on

Returns:
Logical vector of trends by algorithm

Parameter List:

this
-----
Input data frame with the following columns:
dt.ym = Date format with each row being a unique month
dpmo = Calculated complaint rate
count = Number of complaints

mo
---------
Which Date-format date from this$dt.ym to run algorithms on

Returns:
Logical vector of trends by algorithm

Parameter List:

this
-----
Input data frame with the following columns:
dt.ym = Date format with each row being a unique month
dpmo = Calculated complaint rate
count = Number of complaints

mo
---------
Which Date-format date from this$dt.ym to run algorithms on
}
