% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additional.R, R/functions.R, R/qc.R
\name{Uptrend}
\alias{Uptrend}
\alias{Uptrend}
\alias{Uptrend}
\title{Uptrend Slope}
\usage{
Uptrend(inset, eval.date, n.mos = 3)

Uptrend(inset, eval.date, n.mos = 3)

Uptrend(inset, eval.date, n.mos = 3)
}
\description{
Returns:
Slope (dpmo/month) for last n months

Returns:
Slope (dpmo/month) for last n months

Returns:
Slope (dpmo/month) for last n months
}
\details{
Parameter List:

inset
-----
Input data frame with the following columns:
dt.ym = Date format with each row being a unique month
dpmo = Number of complaints

eval.date
---------
Which Date-format date in inset$dt.ym to run algorithm on

n.mos
-----
The number of months to estimate the slope from

Parameter List:

inset
-----
Input data frame with the following columns:
dt.ym = Date format with each row being a unique month
dpmo = Number of complaints

eval.date
---------
Which Date-format date in inset$dt.ym to run algorithm on

n.mos
-----
The number of months to estimate the slope from

Parameter List:

inset
-----
Input data frame with the following columns:
dt.ym = Date format with each row being a unique month
dpmo = Number of complaints

eval.date
---------
Which Date-format date in inset$dt.ym to run algorithm on

n.mos
-----
The number of months to estimate the slope from
}
