% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{CUSUM}
\alias{CUSUM}
\title{CUSUM Signals - 4 Western Electric Rules}
\usage{
CUSUM(inset, eval.date, nsigma = 1.5, zerorate = (2/3))
}
\description{
Returns:
Vector of 4 with a logical of whether the Zone Rules 1-4 were met
}
\details{
Parameter List:

inset
-----
Input data frame with the following columns:
dt.ym = Date format with each row being a unique month
dpmo = Calculated complaint rate

eval.date
---------
Which Date-format date in inset$dt.ym to run algorithm on

nsigma
-----
Number of sigmas for which to detect a mean shift in CUSUM

zerorate
--------
Maximum proportion of months containing zeroes for this algorithm to run
}
