% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{define_algos}
\alias{define_algos}
\title{Define Algorithms to Run}
\usage{
define_algos(algos, paramlist)
}
\arguments{
\item{df}{Required input object of class \code{mdsstat_test}}
}
\value{
1-row data frame (class \code{mdsstat_df}) summarizing the test.
}
\description{
Define any number of algorithms with various parameter settings and save as a
reusable set of instructions.
}
\examples{
test2row(prr(mds_ts[[3]]))
}
