% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additional.R
\name{poisson_rare}
\alias{poisson_rare}
\title{Poisson for Rare Events}
\usage{
poisson_rare(df, ...)
}
\arguments{
\item{df}{Input data frame with the following columns:
\describe{
  \item{time}{Unique times of class \code{Date}}
  \item{event}{Either the event count or rate of class \code{numeric}. If
  \code{df} is of class \code{mds_ts}, this column will be created based on
  the value of the \code{ts_event} parameter.}
}}

\item{eval_period}{A positive integer indicating the number of unique times
counting in reverse chronological order to assess.

Default: \code{NULL} considers all times in \code{df}.}

\item{zero_rate}{Minimum proportion of \code{event}s in \code{df}
(constrained by \code{eval_period}) containing zeroes for this algorithm to
run.

Default: \code{2/3} requires a minimum of 2/3 zeros in \code{event}s in
\code{df}.}

\item{p_rate}{Hypothesized Poisson rate parameter null value at which the
Poisson test is performed (null vs. greater). See details for more.

Default: \code{0.2}}

\item{p_crit}{Critical p-value for the Poisson test..

Default: \code{0.05}}
}
\description{
Test on rare events using an exact test on the Poisson distribution rate
parameter (\code{stats::poisson.test()}).
}
\details{
\code{p_rate} default of \code{0.2} is a suggested null value for
the Poisson rate parameter. However this value is highly advised to be set
based on known priors and/or your specific application.
}
